requirements.txt
selenium==4.25.0
webdriver-manager==4.0.2
python-telegram-bot==13.15
python-dotenv==1.0.1
pytz==2024.2
pandas==2.2.3
requests==2.32.3
urllib3==2.2.3

